requirements.txt
numpy==1.25.0
torchvision==0.15.2
visdom==0.2.4
Pillow==9.4.0
scikit-learn==1.2.2
scipy==1.10.1
matplotlib==3.7.1
setuptools==67.8.0
tqdm==4.65.0
scikit-image==0.20.0
tifffile==2021.7.2
timm~=0.9.2